refactor: move IStepContext into SDK and impl from Core to Execution
