+++ src/Xbehave.2.Core/Xbehave.2.Core.csproj
@@ -70,7 +70,7 @@
     <Compile Include="ExampleAttribute.cs" />
     <Compile Include="Fluent\IStepBuilder.cs" />
     <Compile Include="Fluent\StepBuilder.cs" />
-    <Compile Include="IStepContext.cs" />
+    <Compile Include="Sdk\IStepContext.cs" />
     <Compile Include="ScenarioAttribute.cs" />
     <Compile Include="Sdk\ThreadStaticStepHub.cs" />
     <Compile Include="Sdk\Guard.cs" />
