+++ src/Xbehave.2.Execution.desktop/Xbehave.2.Execution.desktop.csproj
@@ -64,6 +64,7 @@
     <Compile Include="ScenarioInvoker.cs" />
     <Compile Include="ScenarioRunner.cs" />
     <Compile Include="Step.cs" />
+    <Compile Include="StepContext.cs" />
     <Compile Include="StepInvoker.cs" />
     <Compile Include="StepRunner.cs" />
     <Compile Include="TypeInfoExtensions.cs" />
